Attempt cleanup when failing to create secret
If an application requests multiple services, there is a chance that
one handler has applied successfully, and another fails.

In this case, it is important to run cleanup, so that any service
users or similar created by the successful handler is cleaned up
properly, since the secret is not getting saved to the cluster.