Count serviceusers by naming convention (dot vs underscore)
In a branch soon to be merged, we change the convention used for generating
serviceusernames. The new convention uses _ as separator, which is illegal
in other parts of the name. Adding metrics for each convention will help us
track when the new convention has completely replaced the old.

By adding the metrics before we merge the branch, we can verify the counts
from the start.