Reuse existing service users in case of collision
There is a small risk that a collision would happen across clusters, in which case we would get in trouble when one of the secrets is deleted and the service user is deleted with it.
The chances of this happening are small, so we think it is an acceptable risk.

Co-authored-by: Trong Huu Nguyen <trong.huu.nguyen@nav.no>
Co-authored-by: Jan-Kåre Solbakken <jan-kare.solbakken@nav.no>
Co-authored-by: Hans Kristian Flaatten <hans.kristian.flaatten@nav.no>