Allow all @read/@write commands even if they are dangerous.
Of the @dangerous commands, we are mostly concerned with the admin type commands (slave, replicas, shutdown etc.) and not the commands that "change lots" (flushdb for instance).

This way we give admin, remove all the dangerous commands, and make sure to give relevant read and write commands even if they were dangerous.