Add label to select/deselect protected secrets
We might need to select secrets based on labels, and wish to include/exclude protected secrets from the set. When protected status is just an annotation, we can't do that, so make sure to set it as a label as well.
Consider removing the annotation at a later time.

This will not affect older secrets, so until all protected secrets have been rotated, this label will not be available for use.