fix: skip service address cache for team-managed services
Kafka is the only service that is managed by us and rarely has its
addresses changed, which makes sense for cached lookups. It especially
makes sense for Kafka due to its size in Nav.

The other services, OpenSearch and Valkey, may be freely created and
deleted by the teams themselves - which means we either need to watch
for deletion events and accordingly remove the cache entries, or
skip the cache lookups entirely. We choose the latter here.